various updates from updating canonical protobuf JSON mappings
The most salient of which is that flow.DocsAndBytes now uses u32 instead
of u64 for accumulations of count mass.

DocsAndBytes is a protobuf message only within the scope of a single
transaction. Thereafter it's serialized as JSON and reduced using `sum`
annotations, which work with u64. So the impact of this change is minor
so long as no single transaction has more than 4GB of document bytes
processed.

We may want to revisit this, perhaps by allowing coercion with
string-encoded numeric values within the `sum` annotation.

+++ crates/derive/src/pipeline/invocation.rs
@@ -125,11 +125,11 @@ impl Invocation {
                                 .context("failed to parse lambda invocation response")?;
                             let stats = InvokeStats {
                                 output: DocCounter::new(
-                                    parsed.iter().map(Vec::len).sum::<usize>() as u64,
+                                    parsed.iter().map(Vec::len).sum::<usize>() as u32,
                                     // TODO: _technically_ we _could_ subtract some bytes here to
                                     // account for the square brackets used in array encoding. I'm
                                     // not sure whether we should or not.
-                                    data.len() as u64,
+                                    data.len() as u32,
                                 ),
                                 total_duration,
                             };
